requirements.txt
beautifulsoup4==4.10.0
boxx==0.9.11
bpycv==0.2.43
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
fn==0.2.4
fonttools==4.31.2
idna==3.3
imageio==2.16.1
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
minexr==1.0.1
natsort==8.1.0
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
PyOpenGL==3.1.6
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
snakeviz==2.1.1
soupsieve==2.3.1
threadpoolctl==3.1.0
tifffile==2022.3.25
tornado==6.1
urllib3==1.26.9
yacs==0.1.8
zcs==0.1.17
